+++ src/bitmap/bitmap_data_writer.cpp
@@ -38,57 +38,8 @@ namespace Invader {
 
             // Get the data
             std::vector<std::byte> current_bitmap_pixels(reinterpret_cast<const std::byte *>(bitmap_color_plate.pixels.data()), reinterpret_cast<const std::byte *>(bitmap_color_plate.pixels.data() + bitmap_color_plate.pixels.size()));
-
-            // Determine if there is any alpha present
-            enum AlphaType {
-                ALPHA_TYPE_NONE,
-                ALPHA_TYPE_ONE_BIT,
-                ALPHA_TYPE_MULTI_BIT
-            };
-
-            bool alpha_equals_luminosity = true;
-            bool luminosity_set = false;
-
-            AlphaType alpha_present = ALPHA_TYPE_NONE;
             auto *first_pixel = reinterpret_cast<ColorPlatePixel *>(current_bitmap_pixels.data());
-            auto *last_pixel = reinterpret_cast<ColorPlatePixel *>(current_bitmap_pixels.data() + current_bitmap_pixels.size());
-
-            // If we need to do monochrome, check if the alpha equals luminosity
-            if(format == BitmapFormat::BITMAP_FORMAT_MONOCHROME) {
-                for(auto *pixel = first_pixel; pixel < last_pixel; pixel++) {
-                    std::uint8_t luminosity = pixel->convert_to_y8();
-
-                    // First, check if the luminosity is the same as alpha. If not, AY8 is not an option.
-                    if(luminosity != pixel->alpha) {
-                        alpha_equals_luminosity = false;
-
-                        // Next, check if luminosity is not 0xFF. If so, A8 is not an option
-                        if(luminosity != 0xFF) {
-                            luminosity_set = true;
-                        }
-
-                        // Next, check if the alpha is set. If so, A8Y8 or A8 are options. Otherwise, Y8 is what we can do.
-                        if(pixel->alpha != 0xFF) {
-                            alpha_present = ALPHA_TYPE_MULTI_BIT;
-                        }
-                    }
-                }
-            }
-
-            // If we aren't doing monochrome, then we check the bitness of the alpha (1-bit = 0xFF or 0x00, which if we only had that, we can get 5-bit color on 16-bit ARGB instead of 4-bit color)
-            else {
-                for(auto *pixel = first_pixel; pixel < last_pixel; pixel++) {
-                    if(pixel->alpha != 0xFF) {
-                        if(pixel->alpha == 0) {
-                            alpha_present = ALPHA_TYPE_ONE_BIT;
-                        }
-                        else {
-                            alpha_present = ALPHA_TYPE_MULTI_BIT;
-                            break;
-                        }
-                    }
-                }
-            }
+            bitmap.format = BitmapEncode::most_efficient_format(current_bitmap_pixels.data(), bitmap.width, bitmap.height, bitmap.depth, format, bitmap.type);
 
             // Set the format
             bool compressed = (format == BitmapFormat::BITMAP_FORMAT_DXT1 || format == BitmapFormat::BITMAP_FORMAT_DXT3 || format == BitmapFormat::BITMAP_FORMAT_DXT5);
@@ -102,57 +53,8 @@ namespace Invader {
             // Set palettized
             bool palettized = false;
 
-            switch(format) {
-                case BitmapFormat::BITMAP_FORMAT_32_BIT:
-                    bitmap.format = alpha_present == AlphaType::ALPHA_TYPE_NONE ? BitmapDataFormat::BITMAP_DATA_FORMAT_X8R8G8B8 : BitmapDataFormat::BITMAP_DATA_FORMAT_A8R8G8B8;
-                    break;
-                case BitmapFormat::BITMAP_FORMAT_16_BIT:
-                    switch(alpha_present) {
-                        case ALPHA_TYPE_NONE:
-                            bitmap.format = BitmapDataFormat::BITMAP_DATA_FORMAT_R5G6B5;
-                            break;
-                        case ALPHA_TYPE_ONE_BIT:
-                            bitmap.format = BitmapDataFormat::BITMAP_DATA_FORMAT_A1R5G5B5;
-                            break;
-                        case ALPHA_TYPE_MULTI_BIT:
-                            bitmap.format = BitmapDataFormat::BITMAP_DATA_FORMAT_A4R4G4B4;
-                            break;
-                    }
-                    break;
-                case BitmapFormat::BITMAP_FORMAT_MONOCHROME:
-                    if(alpha_equals_luminosity) {
-                        bitmap.format = BitmapDataFormat::BITMAP_DATA_FORMAT_AY8;
-                    }
-                    else if(alpha_present == ALPHA_TYPE_MULTI_BIT) {
-                        if(luminosity_set) {
-                            bitmap.format = BitmapDataFormat::BITMAP_DATA_FORMAT_A8Y8;
-                        }
-                        else {
-                            bitmap.format = BitmapDataFormat::BITMAP_DATA_FORMAT_A8;
-                        }
-                    }
-                    else {
-                        bitmap.format = BitmapDataFormat::BITMAP_DATA_FORMAT_Y8;
-                    }
-                    break;
-
-                case BitmapFormat::BITMAP_FORMAT_DXT1:
-                    bitmap.format = BitmapDataFormat::BITMAP_DATA_FORMAT_DXT1;
-                    break;
-                case BitmapFormat::BITMAP_FORMAT_DXT3:
-                    bitmap.format = alpha_present == AlphaType::ALPHA_TYPE_NONE ? BitmapDataFormat::BITMAP_DATA_FORMAT_DXT1 : BitmapDataFormat::BITMAP_DATA_FORMAT_DXT3;
-                    break;
-                case BitmapFormat::BITMAP_FORMAT_DXT5:
-                    bitmap.format = alpha_present == AlphaType::ALPHA_TYPE_NONE ? BitmapDataFormat::BITMAP_DATA_FORMAT_DXT1 : BitmapDataFormat::BITMAP_DATA_FORMAT_DXT5;
-                    break;
-
-                default:
-                    eprintf_error("Unsupported bitmap format.");
-                    throw InvalidBitmapFormatException();
-            }
-
             // Determine if we should use P8 bump
-            bool should_p8 = usage == BitmapUsage::BITMAP_USAGE_HEIGHT_MAP && palettize;
+            bool should_p8 = (usage == BitmapUsage::BITMAP_USAGE_HEIGHT_MAP || usage == BitmapUsage::BITMAP_USAGE_VECTOR_MAP) && palettize;
             if(should_p8) {
                 compressed = false;
                 bitmap.format = BitmapDataFormat::BITMAP_DATA_FORMAT_P8_BUMP;
@@ -199,7 +101,7 @@ namespace Invader {
 
             #define BYTES_TO_MIB(bytes) (bytes / 1024.0F / 1024.0F)
 
-            oprintf("    Bitmap #%zu: %ux%u, %u mipmap%s, %s%s - %.03f MiB\n", i, scanned_color_plate.bitmaps[i].width, scanned_color_plate.bitmaps[i].height, mipmap_count, mipmap_count == 1 ? "" : "s", bitmap_data_format_name(bitmap.format), (bitmap.format == BitmapDataFormat::BITMAP_DATA_FORMAT_DXT1 && alpha_present != AlphaType::ALPHA_TYPE_NONE) ? " with alpha" : "", BYTES_TO_MIB(total_size));
+            oprintf("    Bitmap #%zu: %ux%u, %u mipmap%s, %s - %.03f MiB\n", i, scanned_color_plate.bitmaps[i].width, scanned_color_plate.bitmaps[i].height, mipmap_count, mipmap_count == 1 ? "" : "s", bitmap_data_format_name(bitmap.format), BYTES_TO_MIB(total_size));
         }
     }
 }